+++ google-cloud-spanner/src/test/java/com/google/cloud/spanner/InstanceAdminGaxTest.java
@@ -44,6 +44,7 @@ import io.grpc.Server;
 import io.grpc.StatusRuntimeException;
 import io.grpc.inprocess.InProcessServerBuilder;
 import java.io.IOException;
+import java.time.Duration;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
@@ -64,7 +65,6 @@ import org.junit.runner.RunWith;
 import org.junit.runners.Parameterized;
 import org.junit.runners.Parameterized.Parameter;
 import org.junit.runners.Parameterized.Parameters;
-import org.threeten.bp.Duration;
 
 @RunWith(Parameterized.class)
 public class InstanceAdminGaxTest {
@@ -221,23 +221,23 @@ public class InstanceAdminGaxTest {
     mockInstanceAdmin.reset();
     RetrySettings retrySettingsWithLowTimeout =
         RetrySettings.newBuilder()
-            .setInitialRetryDelay(Duration.ofMillis(1L))
-            .setMaxRetryDelay(Duration.ofMillis(10L))
-            .setInitialRpcTimeout(Duration.ofMillis(20L))
-            .setMaxRpcTimeout(Duration.ofMillis(200L))
+            .setInitialRetryDelayDuration(Duration.ofMillis(1L))
+            .setMaxRetryDelayDuration(Duration.ofMillis(10L))
+            .setInitialRpcTimeoutDuration(Duration.ofMillis(20L))
+            .setMaxRpcTimeoutDuration(Duration.ofMillis(200L))
             .setRetryDelayMultiplier(1.3d)
             .setMaxAttempts(10)
-            .setTotalTimeout(Duration.ofMillis(20000L))
+            .setTotalTimeoutDuration(Duration.ofMillis(20000L))
             .setJittered(false)
             .build();
     RetrySettings retrySettingsWithHighTimeout =
         RetrySettings.newBuilder()
-            .setInitialRetryDelay(Duration.ofMillis(1L))
-            .setMaxRetryDelay(Duration.ofMillis(1L))
-            .setInitialRpcTimeout(Duration.ofMillis(2000L))
-            .setMaxRpcTimeout(Duration.ofMillis(5000L))
+            .setInitialRetryDelayDuration(Duration.ofMillis(1L))
+            .setMaxRetryDelayDuration(Duration.ofMillis(1L))
+            .setInitialRpcTimeoutDuration(Duration.ofMillis(2000L))
+            .setMaxRpcTimeoutDuration(Duration.ofMillis(5000L))
             .setMaxAttempts(3)
-            .setTotalTimeout(Duration.ofMillis(15000L))
+            .setTotalTimeoutDuration(Duration.ofMillis(15000L))
             .build();
     final RetrySettings retrySettingsToUse =
         exceptionType == ExceptionType.DELAYED